Fixed fetchData
